requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
Pillow==10.2.0
protobuf==4.25.2